renamed a bunch of components for more consistent naming
