+++ src/components/Content/DesktopContent/DesktopContent.tsx
@@ -1,7 +1,7 @@
 import ControlsContent from "components/ControlsContent";
 import Footer from "components/Footer";
 import MainContent from "components/MainContent";
-import MapsContent from "components/MapsContent";
+import MapsContent from "components/MapsPanel";
 import React from "react";
 
 export default function DesktopContent(): JSX.Element {
