--- src/components/MapsContent/index.ts
@@ -1,3 +1,0 @@
-import MapsContent from "./MapsContent";
-
-export default MapsContent;
