+++ src/components/MapsPanel/MiraHq.tsx
