+++ src/components/MapsPanel/index.ts
@@ -1,0 +1,3 @@
+import MapsPanel from "./MapsPanel";
+
+export default MapsPanel;
