--- src/components/Notes/index.tsx
@@ -1,3 +1,0 @@
-import Notes from "./Notes";
-
-export default Notes;
