+++ src/components/NotesPanel/NotesPanel.styles.ts
