+++ src/components/NotesPanel/index.tsx
@@ -1,0 +1,3 @@
+import NotesPanel from "./NotesPanel";
+
+export default NotesPanel;
