+++ src/components/PlayersPanel/PlayersPanel.styles.ts
