+++ src/components/PlayersPanel/index.ts
@@ -1,0 +1,3 @@
+import PlayersPanel from "./PlayersPanel";
+
+export default PlayersPanel;
