--- src/components/PlayersSection/index.ts
@@ -1,3 +1,0 @@
-import PlayersSection from "./PlayersSection";
-
-export default PlayersSection;
