+++ src/components/ScoreControls/index.ts
@@ -1,0 +1,3 @@
+import ScoreControls from "./ScoreControls";
+
+export default ScoreControls;
