--- src/components/Scores/index.ts
@@ -1,3 +1,0 @@
-import Scores from "./Scores";
-
-export default Scores;
