+++ src/components/ScoresPanel/CircularProgressBarScores/CircularProgressBarScores.styles.ts
