+++ src/components/ScoresPanel/CircularProgressBarScores/CircularProgressBarScores.tsx
