+++ src/components/ScoresPanel/CircularProgressBarScores/index.ts
